Update README.md
Hi,

the second argument for prepare-iso.sh should only be the iso name without the ending. The script is adding the ending ".iso" and then generates the iso in the temp folder. When it is finished the iso is copied to the desktop. So no fully qualified path is needed here.

Kind regards,

Marc

P.S.: Thanks for the script. Works like a charm!